+++ src/scriptlets/scriptlets-list.js
@@ -71,6 +71,7 @@ export * from './no-protected-audience';
 export * from './trusted-suppress-native-method';
 export * from './json-prune-xhr-response';
 export * from './trusted-dispatch-event';
+export * from './trusted-replace-outbound-text';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export * from './amazon-apstag';