+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - `isRedirectResourceCompatibleWithAdg()` method to check compatibility of redirect resources with AdGuard
   without needing the full rule text [#420]
+- `trusted-replace-outbound-text` scriptlet [#410]
 - `trusted-dispatch-event` scriptlet [#382]
 
 ### Deprecated
@@ -26,6 +27,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v1.10.25...HEAD
 [#425]: https://github.com/AdguardTeam/Scriptlets/issues/425
 [#420]: https://github.com/AdguardTeam/Scriptlets/issues/420
+[#410]: https://github.com/AdguardTeam/Scriptlets/issues/410
 [#382]: https://github.com/AdguardTeam/Scriptlets/issues/382
 
 ## [v1.10.25] - 2024-03-28